Gmail/Gmail/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
python-dotenv
openai
faiss-cpu
pydantic
redis
apscheduler
python-multipart
PyPDF2
aiofiles